move application into an app directory
