+++ app/__init__.py
@@ -1,0 +1,2 @@
+# flake8: noqa
+from .app import manager
