+++ app/db/__init__.py
@@ -1,0 +1,3 @@
+# flake8:noqa
+from app.db.db import *
+from app.db import init
