+++ app/db/db.py
@@ -8,7 +8,7 @@ migrate = Migrate(db)
 
 def init(app, manager):
     db.init_app(app)
-    migrate.init_app(app, db, 'db/migrations')
+    migrate.init_app(app, db, 'app/db/migrations')
     manager.add_command('db', MigrateCommand)
 
 #    with app.app_context():
