+++ app/model/__init__.py
@@ -1,0 +1,3 @@
+# flake8:noqa
+from app.model.invite import Invite
+from app.model.guest import Guest
