+++ app/model/guest.py
@@ -1,4 +1,4 @@
-from db import db
+from app.db import db
 
 
 class Guest(db.Model):
