+++ app/model/invite.py
@@ -1,4 +1,4 @@
-from db import db
+from app.db import db
 import uuid
