+++ app/views/__init__.py
@@ -1,0 +1,4 @@
+# flake8: noqa
+from app.views.home import Home
+from app.views.list_invites import ListInvites
+from app.views.create_invite import CreateInvite
