+++ app/views/create_invite.mustache
