+++ app/views/list_invites.mustache
