--- db/__init__.py
@@ -1,4 +1,0 @@
-# flake8:noqa
-from .db import db
-from .db import init
-from db import *
