--- model/__init__.py
@@ -1,3 +1,0 @@
-# flake8:noqa
-from .invite import Invite
-from .guest import Guest
