--- views/__init__.py
@@ -1,4 +1,0 @@
-# flake8: noqa
-from .home import Home
-from .list_invites import ListInvites
-from .create_invite import CreateInvite
